requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==3.10.1
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
